% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFuncLib.R
\name{Check_Variable}
\alias{Check_Variable}
\title{##Generate the chromatin interaction score skl.}
\usage{
Check_Variable(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
one of the following:
\describe{
\item{error message}{ when the input data frame is missing any of required variables: \code{chr}, \code{gene_id}, \code{enh_start}, \code{enh_end}, \code{distance}, \code{qk}, \code{lk}, \code{skl}, \code{skq} }
\item{list of variables}{ when the data frame includes all required variable but includes some NA values}
\item{success message}{ when the data frame is ready for use.}
}
}
\description{
##Generate the chromatin interaction score skl.
}
\examples{
####################

#### Import or upload an input data. Here, an example data (Valid_Input) will be used.
data(List_Example_Data)
Valid_Input = List_Example_Data$Valid_Input
head(Valid_Input)
Check_Variable(Valid_Input)

Invalid_Input1 = List_Example_Data$Valid_Input1
head(Invalid_Input1)
Check_Variable(Invalid_Input1)     #### gene_id is missing

Invalid_Input2 = List_Example_Data$Valid_Input2
head(Invalid_Input2)
Check_Variable(Invalid_Input2)     #### data includes some NA values.


#####################


}
\seealso{
\code{List_Example_Data}
}
