% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFuncLib_For_Distance.R
\name{Compute_Accuracy}
\alias{Compute_Accuracy}
\title{## Get two accurcy measures using the adjusted distances between genes and eQTLs obtained from two methods.}
\usage{
Compute_Accuracy(DistVec, Post_Prob1, Post_Prob2, gold_label, bDisp = FALSE)
}
\arguments{
\item{DistVec}{a vector of the original distance between genes and eQTL}

\item{Post_Prob1}{a vector of posterior probabilities of eQTL obtained from method 1}

\item{Post_Prob2}{a vector of posterior probabilities of eQTL obtained from method 2}

\item{gold_label}{a vector of gold labels}

\item{bDisp}{an option to display the intermediate result.}
}
\value{
a list of two 100 x 4 matrices, each of which contains true positive, false positive, false negative, true negative rates after using the adjusted distance obtained by two methods.
}
\description{
## Get two accurcy measures using the adjusted distances between genes and eQTLs obtained from two methods.
}
\examples{
####################


data(List_Gold_Label)
gold_label = List_Modified_Gold_Label$muther_eqtl     ### Use Muther eQTL data

data(DistVec)   ### Use DistVec for this example only. 

data(List_NeQTL)
Prob_NeQTL_qk = List_NeQTL$Prob_NeQTL_qk

data(List_Naive_Bayes)
Prob_Naive_qk = List_Naive_Bayes$Prob_Naive_qk

List_Acc = Compute_Accuracy(DistVec, Prob_NeQTL_qk, Prob_Naive_qk, gold_label, bDisp=FALSE)

Acc_NeQTL = List_Acc$Accuracy_Mat1
Acc_Naive = List_Acc$Accuracy_Mat2

head(Acc_NeQTL)
head(Acc_Naive)


##### Draw the adjusted ROC
Draw_Adjusted_ROC(Acc_NeQTL, Acc_Naive, "NeQTL", "Naive")

}
\seealso{
\code{List_Acc_Example}, \code{Draw_Adjusted_ROC()}
}
