% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFuncLib_For_Distance.R
\name{Dist_SNP_Gene}
\alias{Dist_SNP_Gene}
\title{# Get a distance between putative eQTL SNPs and associated genes.}
\usage{
Dist_SNP_Gene(prob_qk, DistVec, percent)
}
\arguments{
\item{prob_qk}{a vector of posterior probabilities of SNP to be an eQTL}

\item{DistVec}{a vector of distance between genes and SNPs.}

\item{percent}{a percentage usded for thresholding.}
}
\value{
a list of
\describe{
\item{Dist_eQTL_Gene}{a vector of distance between putative eQTL SNPs and genes}
\item{ProbVal}{a probability used to determine eQTL, which is a (1-percent) percentile of prob_qk}
}
}
\description{
# Get a distance between putative eQTL SNPs and associated genes.
}
\examples{
####################


#'
data(List_NeQTL)
Prob_NeQTL_qk = List_NeQTL$Prob_NeQTL_qk
#####################


data(DistVec)   ### Use DistVec for this example only. 
percent = 0.05
lst = Dist_SNP_Gene(Prob_NeQTL_qk, DistVec, percent)
Dist_SNP_Gene_NeQTL = lst$Dist_eQTL_Gene
Threshold_NeQTL = lst$ProbVal

hist(Dist_SNP_Gene_NeQTL, main="Distribution of distances between eQTLs and gene")


}
\seealso{
GetHist2()
}
