% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DistVec}
\alias{DistVec}
\title{#####################}
\format{
A vector of distance between SNP and genes from real data used in this study.
}
\usage{
data(DistVec)
}
\description{
#####################
}
\examples{

data(DistVec)
summary(DistVec)

}
\keyword{datasets}
