% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFuncLib_For_Distance.R
\name{Draw_Adjusted_ROC}
\alias{Draw_Adjusted_ROC}
\title{## Draw the ROC curves of two methods using the adjusted distances.}
\usage{
Draw_Adjusted_ROC(Accuracy_Mat1, Accuracy_Mat2, method_name1, method_name2)
}
\arguments{
\item{Accuracy_Mat1}{a 100 x 4 matrix of the adjusted accuracy measures of method 1.}

\item{Accuracy_Mat2}{a 100 x 4 matrix of the adjusted accuracy measures of method 2.}

\item{method_name1}{the name of method 1. This will be expressed in the plot.}

\item{method_name2}{the name of method 2. This will be expressed in the plot.}

\item{bDisp}{an option to display the intermediate result.}
}
\value{
two ROC curves using the adjusted distance obtained by two methods.
}
\description{
## Draw the ROC curves of two methods using the adjusted distances.
}
\examples{
####################


###### Here we will use already-adjusted accuracy measures we found in our study. Thus, we skip the process of adjusting measures. See the example in "Compute_Accuracy".
data(List_Acc_Example)

Acc_NeQTL = List_Acc_Example$Accuracy_Mat1
Acc_Naive = List_Acc_Example$Accuracy_Mat2

head(Acc_NeQTL)
head(Acc_Naive)

#'Draw_Adjusted_ROC(Acc_NeQTL, Acc_Naive, "NeQTL", "Naive")


}
\seealso{
\code{List_Acc_Example}, \code{Compute_Accuracy}
}
