% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFuncLib_For_Distance.R
\name{GetHist2}
\alias{GetHist2}
\title{## Get two histograms of distances between genes and eQTLs obtained from two methods.}
\usage{
GetHist2(
  Dist_a,
  Dist_b,
  mainstr = "Plot of two histograms",
  legendstr = c("Blue: Method A", "Pink: Method B")
)
}
\arguments{
\item{Dist_a}{a vector of distance between genes and eQTL obtained from the method a (denoted by blue in the histogram)}

\item{Dist_b}{a vector of distance between genes and eQTL obtained from the method b (denoted by pink in the histogram)}

\item{mainstr}{a title of the histogram}

\item{legendstr}{a vector of legend titles for the method a and method b. Default is c("Blue: Method A", "Pink: Method B")}
}
\value{
a plot of two histogram.
}
\description{
## Get two histograms of distances between genes and eQTLs obtained from two methods.
}
\examples{
####################


data(List_Naive_Bayes)
Prob_Naive_qk = List_Naive_Bayes$Prob_Naive_qk

#####Compare it with NeQTL.

data(List_NeQTL)
Prob_NeQTL_qk = List_NeQTL$Prob_NeQTL_qk
#####################



data(DistVec)   ### Use DistVec for this example only. 
percent = 0.05
lst = Dist_SNP_Gene(Prob_NeQTL_qk, DistVec, percent)
Dist_SNP_Gene_NeQTL = lst$Dist_eQTL_Gene
Threshold_NeQTL = lst$ProbVal
lst = Dist_SNP_Gene(Prob_Naive_qk, DistVec, percent)
Dist_SNP_Gene_Naive = lst$Dist_eQTL_Gene
Threshold_Naive = lst$ProbVal


mainstr="Distances between eQTL SNP's and genes"
legendstr=c("Blue: NeQLT", "Pink: Naive")
GetHist2(Dist_SNP_Gene_NeQTL, Dist_SNP_Gene_Naive, mainstr=mainstr, legendstr=legendstr)

}
\seealso{
Dist_SNP_Gene()
}
