% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{List_Example_Data}
\alias{List_Example_Data}
\title{Example of valid input}
\format{
A list containing examples of valid and invalid input data:
\describe{
  \item{Valid_Input}{Example of a valid input data which contains all required varialbe.}
  \item{Invalid_Input1}{Example of an invalid input data.}
  \item{Invalid_Input2}{Another example of an invalid input data.}
}
}
\usage{
data(List_Example_Data)
}
\description{
#####################
}
\examples{

data(List_Example_Data)
Input_Data = List_Example_Data$Valid_Input
}
\keyword{datasets}
