% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{List_Modified_Gold_Label}
\alias{List_Modified_Gold_Label}
\title{Dataset of modified gold labels for sample examples only. For your research, use List_Gold_Label.}
\format{
A list containing modified gold labels (vectors of true or false):
\describe{
  \item{battle_eqtl}{gold label data of eQTL from Battle (2014).}
  \item{muther_eqtl}{gold label data of eQTL from Multiple Tissue Human Expression Resource (Muther) project.}
  \item{geuvadis_eqtl}{gold label data of eQTL from Genetic European Variation in Disease (Geuvadis) project.}
}
}
\usage{
data(List_Modified_Gold_Label)
}
\description{
#####################
}
\examples{

data(List_Modified_Gold_Label)
battle = List_Modified_Gold_Label$battle_eqtl
muther = List_Modified_Gold_Label$muther_eqtl
geuvadis = List_Modified_Gold_Label$geuvadis_eqtl



}
\references{
[1] Battle, A. et al. (2014) Characterizing the genetic basis of transcriptome diversity through RNA-sequencing of 922 individuals. Genome. Res. 24: 14-24
}
\seealso{
List_Gold_Label, Mod_Gold_Label()
}
\keyword{datasets}
