% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{List_Naive_Bayes}
\alias{List_Naive_Bayes}
\title{Sample result obtained from the application of naive Bayesian method, which doesn't use the neighboring relationship.}
\format{
A list of
\describe{
  \item{Prob_Naive_qk}{posterior probabilities for qk=1}
}
}
\usage{
data(List_Naive_Bayes)
}
\description{
#####################
}
\examples{

data(List_Naive_Bayes)
Prob_Naive_qk = List_Naive_Bayes$Prob_Naive_qk

#####Compare it with NeQTL.

data(List_NeQTL)
Prob_NeQTL_qk = List_NeQTL$Prob_NeQTL_qk

#### example of computing the auc using muther gold label
data(List_Modified_Gold_Label)
muther = List_Modified_Gold_Label$muther_eqtl

df = cbind("gold_label"=muther, "NeQTL"=Prob_NeQTL_qk, "Naive"=Prob_Naive_qk)
df = data.frame(df)

library(ROCR)
#### NeQTL's auc
pred  = prediction(df$NeQTL, df$gold_label)
perf  = performance(pred, "tpr", "fpr")
auc_b = performance(pred, 'auc')
auc_b = unlist(slot(auc_b,"y.values"))
auc_b =  round(auc_b, digits=3)
plot(perf, xlim=c(0,1), ylim=c(0,1), col='red', main=paste("ROC curve of eQTL prediction"), lwd = 2)
print(auc_b)


pred  = prediction(df$Naive, df$gold_label)
perf  = performance(pred, "tpr", "fpr")
auc_t = performance(pred, 'auc')
auc_t = unlist(slot(auc_t, "y.values"))
auc_t =  round(auc_t, digits=3)
par(new=TRUE)
plot(perf, xlim=c(0,1), ylim=c(0,1), col='blue', lwd = 2)
legend("bottomright", lwd= c(2,2), col = c("red", "blue"),
      legend = c(paste0("NeQTL, auc=", auc_b), paste0("Naive Bayesian, auc=", auc_t)))

print(auc_t)
}
\seealso{
List_NeQTL
}
\keyword{datasets}
