% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{List_NeQTL}
\alias{List_NeQTL}
\title{Sample result obtained from NeQTL}
\format{
A list of
\describe{
  \item{Prob_NeQTL_qk}{posterior probabilities for qk=1}
}
}
\usage{
data(List_NeQTL)
}
\description{
#####################
}
\examples{

data(List_NeQTL)
Prob_NeQTL_qk = List_NeQTL$Prob_NeQTL_qk

#### example of computing the auc using battle gold label
data(List_Modified_Gold_Label)
battle = List_Modified_Gold_Label$battle_eqtl

df = cbind("gold_label"=battle, "NeQTL"=Prob_NeQTL_qk)
df = data.frame(df)

library(ROCR)
#### NeQTL's auc
pred  = prediction(df$NeQTL, df$gold_label)
perf  = performance(pred, "tpr", "fpr")
auc_b = performance(pred, 'auc')
auc_b = unlist(slot(auc_b,"y.values"))
auc_b =  round(auc_b, digits=3)
plot(perf, xlim=c(0,1), ylim=c(0,1), col='red', main=paste("ROC curve of eQTL prediction"), lwd = 2)
print(auc_b)



}
\seealso{
List_Naive_Bayes
}
\keyword{datasets}
