% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeQTL_Main.R
\name{Mod_Gold_Label}
\alias{Mod_Gold_Label}
\title{## Adapt the gold label dataset to practitioner's needs. Use this function and create a gold label data when your research data doesn't exactly match the reference gold label dataset.}
\usage{
Mod_Gold_Label(Input_Data, original_gold_label)
}
\arguments{
\item{Input_Data}{a data frame that should contain
\describe{
\item{gene_id}{ensembel gene id.}
\item{chr}{a chromosome where the SNP reside.}
\item{gv_position}{a location of the SNP.}
}}

\item{original_gold_label}{a data frame of any gold label reference that include 
\describe{
\item{gene_id}{ensembel gene id.}
\item{chr}{a chromosome where the SNP reside.}
\item{gv_position}{a location of the SNP.}
}}
}
\value{
a vector of gold label.
}
\description{
## Adapt the gold label dataset to practitioner's needs. Use this function and create a gold label data when your research data doesn't exactly match the reference gold label dataset.
}
\examples{
####################

####Import or upload an input data. Here, an example data (Valid_Input) will be used.
data(List_Example_Data)
Input_Data = List_Example_Data$Valid_Input
head(Input_Data)   

data(List_Gold_Label)       ##### Original reference data                     
battle_df = List_Gold_Label$battle_eqtl
gold_label = Mod_Gold_Label(Input_Data, battle_df)   #### Obtain the modified gold lable data.
#####################
}
\references{
[1] Battle, A. et al. (2014) Characterizing the genetic basis of transcriptome diversity through RNA-sequencing of 922 individuals. Genome. Res. 24: 14-24
}
\seealso{
List_Gold_Label, List_Modified_Gold_Label
}
