\name{broman-internal}
\alias{smoothall}

\title{Internal neqtl functions}

\description{
  Internal neqtl functions.
}

\usage{
smoothall(themax, thechr, thepos, win=5)
}

\details{
  These are generally not to be called by the user.
}

\author{Karl W Broman \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
