% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeQTL_Main.R
\name{Run_NeQTL2}
\alias{Run_NeQTL2}
\title{##Performs Bayesian inference for eQTL analysis}
\usage{
Run_NeQTL2(
  Input_Data,
  qR = 0.1,
  lR = 0.1,
  b_gen_skl = FALSE,
  b_use_lNk = FALSE,
  total_iter = 100,
  bDisp = FALSE
)
}
\arguments{
\item{Input_Data}{a data frame that contains
\describe{
\item{chr}{a chromosome where SNP is located.}
\item{gene_id}{an ensemble gene id.}
\item{enh_start}{an enhancer starting point.}
\item{enh_end}{an enhancer ending point.}
\item{qk}{the eQTL status (qk) of the kth SNP: qk=1 if the SNP is an eQTL and 0 otherwise.}
\item{lk}{the chromatin interaction status (lk) of the kth SNP: lk=1 if the SNP has a chromatin interaction and 0 otherwise.}
\item{skq}{the eQTL score.}
\item{skl}{the chromatin interaction score.}
}}

\item{qR}{threshold value for eQTL. Default is 0.1.}

\item{lR}{threshold value for chromatin interaction. Default is 0.1.}

\item{b_gen_skl}{option for generating the chromatin interaction scores (skl) internally. Default is FALSE. To use this option, the input data should include "Gene_Activity", "Enh_Activity", and "Correlation": see Generate_skl(). Using option, a better skl based on the findings will be generated.}

\item{b_use_lNk}{option for incorporating a neighboring relationship for chromatin interaction. Default is FALSE.}

\item{total_iter}{number of iterations. Default is set at 100. Larger than the default is recommended.}

\item{bDisp}{option for displaying the iteration. Default is FALSE.}
}
\value{
list of the following values:
\describe{
\item{lk_index}{Indice of chromatin interactions.}
\item{qk_index}{Indice of eQTL SNPs.}
\item{qNk_index}{Indice of neighboring eQTLs.}
\item{prob_qk1_all}{Probability of a SNP being eQTL.}
\item{prob_lk1_all}{Probablity of a chromatin interaction.}
\item{prob_qNk_qk_0}{Conditional probability of neighboring eQTL SNPs given that it is not an eQTL SNP.}
\item{prob_qNk_qk_1}{Conditional probability of neighboring eQTL SNPs given that it is an eQTL SNP.}
\item{skq_qk0lk0}{eQTL score given no eQTL SNP (qk=0) and no chromatin interaction (lk=0).}
\item{skq_qk0lk1}{eQTL score given no eQTL SNP (qk=0) and chromatin interaction (lk=1).}
\item{skq_qk1lk0}{eQTL score given eQTL SNP (qk=1) and no chromatin interaction (lk=0).}
\item{skq_qk1lk1}{eQTL score given eQTL SNP (qk=1) and chromatin interaction (lk=1).}
\item{skl_qk0lk0}{chromatin interaction score given no eQTL SNP (qk=0) and no chromatin interaction (lk=0).}
\item{skl_qk0lk1}{chromatin interaction score given no eQTL SNP (qk=0) and chromatin interaction (lk=1).}
\item{skl_qk1lk0}{chromatin interaction score given eQTL SNP (qk=1) and no chromatin interaction (lk=0).}
\item{skl_qk1lk1}{chromatin interaction score given eQTL SNP (qk=1) and chromatin interaction (lk=1).}

}
}
\description{
##Performs Bayesian inference for eQTL analysis
}
\examples{
####################

## Import or upload an input data. Here, an example data (Valid_Input) will be used.
data(List_Example_Data)
Input_Data = List_Example_Data$Valid_Input   #### This is a short version of the real data used in this study.
head(Input_Data)
dim(Input_Data)
Final_lst = Run_NeQTL2(Input_Data, qR=0.1, lR=0.1, b_gen_skl=FALSE, 
b_use_lNk=FALSE, total_iter=10, bDisp=FALSE)

#####################


}
